계좌 리스트 조회 및 대표 계좌 변경

See merge request s09-fintech-finance-sub2/S09P22A108!85

Former-commit-id: f7050293c216553f7642ce1bb9e11e46ca47e4f9

+++ backend/src/main/java/com/sss/tally/api/calculate/dto/CalculateDto.java
@@ -146,7 +146,7 @@ public class CalculateDto {
 
 		private int myAmount;
 
-		private Long allAmount;
+		private Double allAmount;
 
 		private LocalDateTime paymentDate;
